keda/pyproject.toml
[tool.poetry]
name = "keda-restart"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.8"
loki-client = "^0.0.2"
requests = "^2.28.2"
kubernetes = "^26.1.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
